backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
python-multipart>=0.0.12
httpx>=0.28.0
python-dotenv>=1.0.0
deep-translator>=1.11.4
pysrt>=1.1.2
yt-dlp>=2024.12.0
